backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-environ>=0.11.2
psycopg2-binary>=2.9
gunicorn>=22.0
whitenoise>=6.7
